Move utility classes to the utils package
