+++ lib/src/main/java/org/sonarsource/scanner/lib/internal/IsolatedLauncherFactory.java
@@ -33,6 +33,7 @@ import org.sonarsource.scanner.lib.internal.batch.IsolatedLauncher;
 import org.sonarsource.scanner.lib.internal.cache.CachedFile;
 import org.sonarsource.scanner.lib.internal.cache.FileCache;
 import org.sonarsource.scanner.lib.internal.http.ScannerHttpClient;
+import org.sonarsource.scanner.lib.internal.util.TempCleaning;
 
 public class IsolatedLauncherFactory {
