+++ lib/src/main/java/org/sonarsource/scanner/lib/internal/JavaRunnerFactory.java
@@ -42,19 +42,20 @@ import javax.annotation.Nullable;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.sonarsource.scanner.lib.System2;
 import org.sonarsource.scanner.lib.internal.cache.CachedFile;
 import org.sonarsource.scanner.lib.internal.cache.FileCache;
 import org.sonarsource.scanner.lib.internal.cache.HashMismatchException;
 import org.sonarsource.scanner.lib.internal.http.ScannerHttpClient;
 import org.sonarsource.scanner.lib.internal.util.CompressionUtils;
+import org.sonarsource.scanner.lib.internal.util.ProcessWrapperFactory;
+import org.sonarsource.scanner.lib.internal.util.System2;
 
 import static java.lang.String.format;
 import static org.sonarsource.scanner.lib.ScannerProperties.JAVA_EXECUTABLE_PATH;
 import static org.sonarsource.scanner.lib.ScannerProperties.SCANNER_ARCH;
 import static org.sonarsource.scanner.lib.ScannerProperties.SCANNER_OS;
 import static org.sonarsource.scanner.lib.ScannerProperties.SKIP_JRE_PROVISIONING;
-import static org.sonarsource.scanner.lib.Utils.deleteQuietly;
+import static org.sonarsource.scanner.lib.internal.util.Utils.deleteQuietly;
 
 public class JavaRunnerFactory {
