+++ lib/src/main/java/org/sonarsource/scanner/lib/internal/ScannerEngineLauncherFactory.java
@@ -28,11 +28,12 @@ import javax.annotation.Nullable;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.sonarsource.scanner.lib.System2;
 import org.sonarsource.scanner.lib.internal.cache.CachedFile;
 import org.sonarsource.scanner.lib.internal.cache.FileCache;
 import org.sonarsource.scanner.lib.internal.cache.HashMismatchException;
 import org.sonarsource.scanner.lib.internal.http.ScannerHttpClient;
+import org.sonarsource.scanner.lib.internal.util.ProcessWrapperFactory;
+import org.sonarsource.scanner.lib.internal.util.System2;
 
 public class ScannerEngineLauncherFactory {
