+++ lib/src/main/java/org/sonarsource/scanner/lib/internal/util/ArchResolver.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.scanner.lib.internal;
+package org.sonarsource.scanner.lib.internal.util;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
@@ -26,7 +26,6 @@ import java.util.Optional;
 import org.apache.commons.lang3.SystemUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.sonarsource.scanner.lib.System2;
 
 public class ArchResolver {
