+++ lib/src/test/java/org/sonarsource/scanner/lib/internal/util/ArchResolverTest.java
@@ -17,14 +17,13 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.scanner.lib.internal;
+package org.sonarsource.scanner.lib.internal.util;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.slf4j.event.Level;
-import org.sonarsource.scanner.lib.System2;
 import testutils.LogTester;
 
 import static org.assertj.core.api.Assertions.assertThat;
